[DESIGN]: 웹 공통 컴포넌트 수정 (#61)
* 🎨 design: profile 컴포넌트 반응형 추가

* 🔨 chore: img 폴더 위치 변경

* 🎨 design: Img 컴포넌트 size 숫자로 변경

* 🎨 design: Ani 컴포넌트 반응형 추가

* 🎨 design: Badge컴포넌트 반응형 추가
