+++ src/components/common/ani/Ani.style.ts
@@ -1,15 +1,18 @@
 import styled from 'styled-components';
-type AniContainerProps = {
-  size: number;
-};
-export const AniContainer = styled.div<AniContainerProps>`
-  width: ${({ size }) => `${size}px`};
+
+export const AniContainer = styled.div`
+  width: 158px;
   aspect-ratio: 1/1;
   background-color: var(--blue-200);
   overflow: hidden;
   display: flex;
   align-items: center;
   justify-content: center;
+
+  //모바일
+  @media (max-width: 768px) {
+    width: 130px;
+  }
 `;
 export const Ani = styled.img`
   width: 100%;
