+++ src/components/common/ani/Ani.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
 import * as S from './Ani.style';
-
 type AniProps = {
-  size?: number;
   imageUrl?: string;
 };
-const Ani: React.FC<AniProps> = ({ size = 130, imageUrl }) => {
+const Ani: React.FC<AniProps> = ({ imageUrl }) => {
   return (
-    <S.AniContainer size={size}>
+    <S.AniContainer>
       {imageUrl && <S.Ani src={imageUrl} alt="애니메이션 이미지" />}
     </S.AniContainer>
   );
