+++ src/components/common/badge/Badge.style.ts
@@ -8,4 +8,9 @@ export const Badge = styled.div`
   border-radius: var(--radius-8);
   display: inline-block;
   font-weight: 500;
+
+  //모바일
+  @media (max-width: 768px) {
+    padding: var(--spacing-8);
+  }
 `;
