+++ src/components/common/img/Img.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import * as S from './Img.style';
 
 type ImgProps = {
-  size: 'xs' | 's' | 'm' | 'l' | 'xl';
+  size: 294 | 168 | 140 | 128 | 100 | 80 | 68;
   imageUrl?: string;
 };
