+++ src/components/common/profile/Profile.style.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
-export const ProfileContainer = styled.div<{ size: number }>`
-  width: ${({ size }) => size}px;
-  height: ${({ size }) => size}px;
+export const ProfileContainer = styled.div<{ size: 'M' | 'S' }>`
+  width: ${({ size }) => (size === 'M' ? 52 : 48)}px;
+  aspect-ratio: 1/1;
   border-radius: 50%;
   background-color: var(--blue-200);
   overflow: hidden;
@@ -11,6 +11,10 @@ export const ProfileContainer = styled.div<{ size: number }>`
   justify-content: center;
   position: relative;
   display: inline-block;
+
+  @media (max-width: 768px) {
+    width: ${({ size }) => (size === 'M' ? 40 : 36)}px;
+  }
 `;
 
 export const ProfileImage = styled.img`
