+++ src/components/common/profile/Profile.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import * as S from './Profile.style';
 
 type ProfileProps = {
-  size?: number;
+  size?: 'M' | 'S';
   imageUrl?: string;
 };
 
-const Profile: React.FC<ProfileProps> = ({ size = 40, imageUrl }) => {
+const Profile: React.FC<ProfileProps> = ({ size = 'M', imageUrl }) => {
   return (
     <S.ProfileContainer size={size}>
       {imageUrl && <S.ProfileImage src={imageUrl} alt="프로필 이미지" />}
